Modify `LAST_SEEN` logic to handle a duplicate notification edge case.
During testing today, I noticed that the `PostingDate` in Apple's JSON
manifest had unexpectedly changed for iOS 15.4.1. That update was
originally released on 2022-03-31 (with a `PostingDate` to match)
but the new `PostingDate` was set to 2022-04-11.

The previous logic was to store a date in the `LAST_SEEN` file and to
pull versions from the PMV manifest that had a more recent posting date.
That logic would have resulted in duplicate notifications being sent for
15.4.1 because the `PostingDate` changed after the initial release.

This didn't end up happening for any alpha users because the timing just
so happened to perfectly coincide with the switch to a new server :)

The `LAST_SEEN` file now stores the latest *version* that was last seen
(instead of its release date) and a version-aware comparison replaces
the date check that was happening before. This should be more resilient
to any future `PostingDate` shenanigans, though it remains a mystery why
that value ever changed in the first place.